Updated bare-metal provisioning framework.
Part 4 of 6: blueprint general-bare-metal-provisioning-framework.

Bare-metal provisioning is extended to use a MySQL and
to support various plug-ins such as PXE.
Driver and domain management codes are updated.
This patch includes only the tilera(non-PXE) back-end.
The documents describe how to install and configure bare-metal
nova-compute to support non-PXE (tilera) back-end nodes.
It includes the packages needed for installation, how to create
a bare-metal instance type, how to register bare-metal nodes and NIC,
and how to run an instance. Nova flags for the configuration of
bare-metal nova-compute are described.
Now one bare-metal driver initially registers multiple entries of bare-metal
nodes. It periodically reports the capabilities of the multiple bare-metal
nodes and reports it as a list of capabilities.

Change-Id: I55617a8da52d20d4df727b8bbde8e5f72d3bf130
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>
DocImpact: adds bare-metal-tilera docs