Added pxe back-end bare-metal.
Part 5 of 6: blueprint general-bare-metal-provisioning-framework.

This patch includes only PXE back-end with IPMI.
The documents describe how to install and configure bare-metal
nova-compute to support PXE back-end nodes.
It includes the packages needed for installation, how to create
a bare-metal instance type, how to register bare-metal nodes and NIC,
and how to run an instance. Nova flags for the configuration of
bare-metal nova-compute are described.

Change-Id: I6358e5ede094b4d887d15913b19f877169a0f99c
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>
DocImpact: adds bare-metal-pxe docs