Updated bare-metal provisioning framework.
Part 4 of 7: blueprint general-bare-metal-provisioning-framework.

New Baremetal driver is implemented in this patch.

This patch does not include Tilera or PXE back-ends, which will be
provided by subsequent patches.

With this driver, nova compute registers multiple entries of baremetal
nodes. It periodically updates the capabilities of the multiple
bare-metal nodes and reports it as a list of capabilities.

Change-Id: I55617a8da52d20d4df727b8bbde8e5f72d3bf130
Co-authored-by: Mikyung Kang <mkkang@isi.edu>
Co-authored-by: David Kang <dkang@isi.edu>
Co-authored-by: Ken Igarashi <igarashik@nttdocomo.co.jp>
Co-authored-by: Arata Notsu <notsu@virtualtech.jp>
Co-authored-by: Chris Krelle <NobodyCam@gmail.com>
Co-authored-by: Devananda van der Veen <devananda.vdv@gmail.com>